feat: simplify and remove redundant files
